requirements.txt
Flask==3.1.2
PyMySQL==1.1.2
Werkzeug==3.1.3
Jinja2==3.1.6
PyJWT==2.10.1
argon2-cffi==25.1.0
python-dotenv==1.2.1